requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.7
fonttools==4.56.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.9.3
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
